requirements.txt
# pip install -r requirements.txt
supervisely==6.1.51

# base ----------------------------------------
PyYAML==5.4.1
tensorboard==2.4.1
tqdm==4.56.0

# plotting ------------------------------------
seaborn==0.11.1

# export --------------------------------------
# coremltools==4.0
# onnx>=1.8.0
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop==0.0.31-2005241907  # FLOPS computation
pycocotools==2.0.2  # COCO mAP
